Introduce ManualFlowController
