+++ lib/membrane/core/element/manual_flow_controller.ex
@@ -1,4 +1,4 @@
-defmodule Membrane.Core.Element.DemandController.Manual do
+defmodule Membrane.Core.Element.ManualFlowController do
   @moduledoc false
 
   # Module handling demands requested on output pads.
@@ -9,6 +9,7 @@ defmodule Membrane.Core.Element.DemandController.Manual do
     ActionHandler,
     BufferController,
     CallbackContext,
+    DemandController,
     EventController,
     InputQueue,
     State,
@@ -47,7 +48,7 @@ defmodule Membrane.Core.Element.DemandController.Manual do
 
   defp do_handle_redemand(pad_ref, state) do
     state = %{state | delay_demands?: true}
-    state = exec_handle_demand(pad_ref, state)
+    state = DemandController.exec_handle_demand(pad_ref, state)
     %{state | delay_demands?: false}
   end
 
@@ -100,11 +101,7 @@ defmodule Membrane.Core.Element.DemandController.Manual do
     %State{state | delay_demands?: false}
   end
 
-  @spec update_demand(
-          Pad.ref(),
-          non_neg_integer() | (non_neg_integer() -> non_neg_integer()),
-          State.t()
-        ) :: State.t()
+  @spec update_demand(Pad.ref(), non_neg_integer() | (non_neg_integer() -> non_neg_integer()), State.t()) :: State.t()
   def update_demand(pad_ref, size, state) when is_integer(size) do
     PadModel.set_data!(state, pad_ref, :manual_demand_size, size)
   end
@@ -205,52 +202,5 @@ defmodule Membrane.Core.Element.DemandController.Manual do
       PadModel.update_data!(state, pad_ref, :manual_demand_size, &(&1 - outbound_metric_buf_size))
 
     BufferController.exec_buffer_callback(pad_ref, buffers, state)
-  end
-
-  @spec exec_handle_demand(Pad.ref(), State.t()) :: State.t()
-  defp exec_handle_demand(pad_ref, state) do
-    with {:ok, pad_data} <- PadModel.get_data(state, pad_ref),
-         true <- exec_handle_demand?(pad_data) do
-      do_exec_handle_demand(pad_data, state)
-    else
-      _other -> state
-    end
-  end
-
-  @spec do_exec_handle_demand(PadData.t(), State.t()) :: State.t()
-  defp do_exec_handle_demand(pad_data, state) do
-    context = &CallbackContext.from_state(&1, incoming_demand: pad_data.incoming_demand)
-
-    CallbackHandler.exec_and_handle_callback(
-      :handle_demand,
-      ActionHandler,
-      %{
-        split_continuation_arbiter: &exec_handle_demand?(PadModel.get_data!(&1, pad_data.ref)),
-        context: context
-      },
-      [pad_data.ref, pad_data.demand, pad_data.demand_unit],
-      state
-    )
-  end
-
-  defp exec_handle_demand?(%{end_of_stream?: true}) do
-    Membrane.Logger.debug_verbose("""
-    Demand controller: not executing handle_demand as :end_of_stream action has already been returned
-    """)
-
-    false
-  end
-
-  defp exec_handle_demand?(%{demand: demand}) when demand <= 0 do
-    Membrane.Logger.debug_verbose("""
-    Demand controller: not executing handle_demand as demand is not greater than 0,
-    demand: #{inspect(demand)}
-    """)
-
-    false
-  end
-
-  defp exec_handle_demand?(_pad_data) do
-    true
   end
 end
